requirements.txt
pillow
requests
python-dotenv